requirements.txt
detectron2.egg==info
matplotlib==3.8.4
matplotlib==3.7.5
numpy==1.24.3
Pillow==10.4.0
scikit_learn==1.4.2
scikit_learn==1.3.2
seaborn==0.13.2
torch==2.3.1
tqdm==4.66.4
